test(diffusion): group vllm model config coverage
Move the quantization dtype coverage next to the other DiffusionVllmModelConfig test and remove the worker-test-only helper import.
